+++ src/main/java/fr/insy2s/utils/wrapper/WrapperListeDepense.java
@@ -1,5 +1,6 @@
 package fr.insy2s.utils.wrapper;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class WrapperListeDepense {
@@ -10,7 +11,7 @@ public class WrapperListeDepense {
 
     private LocalDate date;
 
-    private Double prix;
+    private BigDecimal prix;
 
     private String nomFournisseur;
 
@@ -40,11 +41,11 @@ public class WrapperListeDepense {
         this.date = date;
     }
 
-    public Double getPrix() {
+    public BigDecimal getPrix() {
         return prix;
     }
 
-    public void setPrix(Double prix) {
+    public void setPrix(BigDecimal prix) {
         this.prix = prix;
     }
 
@@ -64,7 +65,7 @@ public class WrapperListeDepense {
         this.etatDepense = etatDepense;
     }
 
-    public WrapperListeDepense(Long id, Long numero, LocalDate date, Double prix, String nomFournisseur, String etatDepense) {
+    public WrapperListeDepense(Long id, Long numero, LocalDate date, BigDecimal prix, String nomFournisseur, String etatDepense) {
         this.id = id;
         this.numero = numero;
         this.date = date;
